% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_meddra.R
\name{join_meddra}
\alias{join_meddra}
\title{Combine together all of the MedDRA terms into a single data.frame}
\usage{
join_meddra(data)
}
\arguments{
\item{data}{MedDRA source data from \code{read_meddra()}}
}
\value{
A data.frame with the "soc_code", "soc_name", "hglt
}
\description{
Combine together all of the MedDRA terms into a single data.frame
}
