% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_meddra.R
\name{read_meddra}
\alias{read_meddra}
\title{Read MedDRA datasets from the source MedDRA datasets}
\usage{
read_meddra(directory)
}
\arguments{
\item{directory}{the directory containing the MedAscii and SeqAscii
directories}
}
\value{
A list of data.frames
}
\description{
Read MedDRA datasets from the source MedDRA datasets
}
